# api-gateway/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/reservation_order": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new reservation order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation_order"
                ],
                "summary": "Create reservation order",
                "parameters": [
                    {
                        "description": "Reservation order data",
                        "name": "reservation_order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation.ReservationOrderUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation.ReservationOrderRes"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reservation_order/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a reservation order by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation_order"
                ],
                "summary": "Get reservation order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation.ReservationOrderRes"
                        }
                    },
                    "400": {
                        "description": "Invalid reservation order ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing reservation order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation_order"
                ],
                "summary": "Update reservation order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated reservation order data",
                        "name": "reservation_order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reservation.ReservationOrderUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation.ReservationOrderUpdate"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Reservation order not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a reservation order by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation_order"
                ],
                "summary": "Delete reservation order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation order deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid reservation order ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Reservation order not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reservation_orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all reservation orders",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation_order"
                ],
                "summary": "Get all reservation orders",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reservation.GetAllReservationOrderRes"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "reservation.GetAllReservationOrderRes": {
            "type": "object",
            "properties": {
                "reservation_order": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reservation.ReservationOrderRes"
                    }
                }
            }
        },
...
        "reservation.MenuRes": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "restaurant": {
                    "$ref": "#/definitions/reservation.Restaurant"
                }
            }
        },
...
        "reservation.ReservationOrderRes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "menu_item": {
                    "$ref": "#/definitions/reservation.MenuRes"
                },
                "quantity": {
                    "type": "string"
                },
                "reservation": {
                    "$ref": "#/definitions/reservation.ReservationRes"
                }
            }
        },
        "reservation.ReservationOrderUpdate": {
            "type": "object",
            "properties": {
                "menu_item_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "reservation_id": {
                    "type": "string"
                }
            }
        },
...
        "reservation.ReservationRes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "reservation_time": {
                    "type": "string"
                },
                "restaurant": {
                    "$ref": "#/definitions/reservation.Restaurant"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/reservation.UserResp"
                }
            }
        },
        "reservation.Restaurant": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
        "reservation.UserResp": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}